split basefragment and state saving
