+++ navigation-base/src/main/java/ru/touchin/roboswag/navigation_base/activities/BaseActivity.kt
@@ -29,7 +29,7 @@ import androidx.appcompat.app.AppCompatActivity
 import ru.touchin.roboswag.core.log.Lc
 import ru.touchin.roboswag.core.log.LcGroup
 import ru.touchin.roboswag.navigation_base.fragments.LifecycleLoggingObserver
-import ru.touchin.roboswag.navigation_viewcontroller.keyboard_resizeable.KeyboardBehaviorDetector
+import ru.touchin.roboswag.navigation_base.keyboard_resizeable.KeyboardBehaviorDetector
 
 /**
  * Created by Gavriil Sitnikov on 08/03/2016.
